refactor(utils): extract URL normalization to a shared utility
