+++ components/SideBar.jsx
@@ -3,6 +3,7 @@ import SideBar from '@node-core/ui-components/Containers/Sidebar';
 import { major } from 'semver';
 import { sidebar } from '#theme/local/site';
 import { version } from '#theme/config';
+import { toPublicLink } from '../utils/helpers/urls.mjs';
 import versions from '../versions.json' with { type: 'json' };
 
 /** @param {string} url */
@@ -15,11 +16,6 @@ const isVersioned = Array.isArray(sidebar);
 
 // Versioned pages expose a version relative path, we prefix it so it matches the absolute sidebar links.
 const versionBase = isVersioned ? `/docs/api/v${version.major}.x` : '';
-
-const pathnameFor = path => {
-  const clean = path.replace(/\/index$/, '').replace(/^\//, '');
-  return (clean ? `${versionBase}/${clean}` : versionBase) || '/';
-};
 
 const groupsFor = path => {
   if (isVersioned) return sidebar;
@@ -39,7 +35,7 @@ const currentVersion = isVersioned ? versionBase : undefined;
 /** Docs sidebar, plus a webpack version picker on the versioned API docs. */
 export default ({ metadata }) => (
   <SideBar
-    pathname={pathnameFor(metadata.path)}
+    pathname={toPublicLink(metadata.path, versionBase)}
     groups={groupsFor(metadata.path)}
     onSelect={redirect}
     as={PrefetchLink}
